Receive lat and lng from CourseSearchActivity

+++ app/src/main/java/com/mojilab/moji/data/LocationData.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 public class LocationData {
     public String mainAddress;
     public String subAddress;
-    public float lat;
-    public float lng;
+    public double lat;
+    public double lng;
 
     public LocationData(){
         this.mainAddress = mainAddress;
@@ -15,7 +15,7 @@ public class LocationData {
         this.lng = lng;
     }
 
-    public LocationData(String mainAddress, String subAddress, float lat, float lng){
+    public LocationData(String mainAddress, String subAddress, double lat, double lng){
         this.mainAddress = mainAddress;
         this.subAddress = subAddress;
         this.lat = lat;
